requirements.txt
# MCP Server dependencies
mcp>=0.9.0

# EPUB processing
ebooklib>=0.18
beautifulsoup4>=4.12
lxml>=4.9.0

# Vector storage
chromadb>=0.4.0

# OpenRouter API (OpenAI compatible)
openai>=1.0.0

# HTTP requests
requests>=2.28.0

# Environment variables
python-dotenv>=1.0.0

# Token counting
tiktoken>=0.5.0

# Progress tracking
tqdm>=4.65.0
